CheafTest/CheafTest/obj/Release/Package/PackageTmp/requirements.txt
django >= 2.1.6, < 3.0 
djangorestframework >= 3.10.1
